backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10
pydantic==2.10.3
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.20
httpx==0.28.1
openai==1.57.4
reportlab==4.2.5
pillow==11.0.0
apscheduler==3.10.4
python-dotenv==1.0.1
email-validator==2.2.0
slowapi==0.1.9
python-dateutil==2.9.0
aiofiles==24.1.0
